+++ src/pygeoprocessing/extensions/ManagedRaster.h
@@ -729,6 +729,8 @@ public:
   // Increments the pointer to the next neighbor
   virtual void next() {
     long xj, yj, flow;
+    delete this->m_ptr;
+    this->m_ptr = nullptr;
     if (i == 8) {
       m_ptr = &endVal;
       return;
@@ -738,6 +740,12 @@ public:
     flow = (pixel.val >> (i * 4)) & 0xF;
     m_ptr = new NeighborTuple(i, xj, yj, static_cast<float>(flow));
     i++;
+  }
+
+  ~NeighborIterator() {
+    if (this->m_ptr != &endVal) {
+      delete this->m_ptr;
+    }
   }
 };